Remove depreciated and unused files
