+++ package.json
@@ -21,7 +21,7 @@
     "@jsdoc/cli": "^0.3.1",
     "bcrypt": "^5.1.1",
     "chalk": "^4.1.2",
-    "discord.js": "^14.11.0",
+    "discord.js": "^14.13.0",
     "dotenv": "^16.3.1",
     "express": "^4.18.2",
     "mongoose": "^7.4.0",
