--- src/events/buttons/buttonInteraction.txt
@@ -1,56 +1,0 @@
-// REMOVED TEMPORARILY TO STOP MEMORY LEAKS.
-// THIS FILE IS REDUNDANT ANYWAY.
-// WHO PUT THIS HERE????
-
-Move slash commands to a new fileconst {sendEmbed, EmbedType} = require("../../functions/utilities/embedUtils");
-const {ActionRowBuilder, ButtonBuilder} = require("discord.js");
-
-module.exports = {
-    name: 'interactionCreate',
-    async execute(client) {
-
-        client.on('interactionCreate', async interaction => {
-
-            if (!interaction.isButton()) return;
-            if (interaction.customId.startsWith('leaderboard')) {
-
-                await interaction.deferUpdate();
-                const {getLeaderboard, calculateXPUntilLevel} = require('../../functions/utilities/levelUtils');
-
-                let leaderboardPage = interaction.customId.split('-')[1];
-                let leaderboard;
-
-                if (interaction.customId.split('-')[0] === 'leaderboard_previous') {
-                    leaderboard = await getLeaderboard(interaction.guild.id, leaderboardPage - 1);
-                    leaderboardPage--;
-                } else if (interaction.customId.split('-')[0] === 'leaderboard_next') {
-                    leaderboard = await getLeaderboard(interaction.guild.id, leaderboardPage + 1);
-                    leaderboardPage++;
-                }
-
-                let leaderboardString = ``;
-                for (let i = 0; i < leaderboard.length; i++) {
-                    const {userID, level, xp} = leaderboard[i];
-                    leaderboardString += `${i + 1}. **${i === 0 ? `:trophy: ` : ``}<@${userID}>** - Level ${level} (${xp} XP)\n`;
-                }
-
-                const leaderboardNavigation = new ActionRowBuilder()
-                    .addComponents(
-                        new ButtonBuilder()
-                            .setCustomId(`leaderboard_previous-${leaderboardPage}`)
-                            .setLabel('Previous')
-                            .setDisabled(leaderboardPage === 1)
-                            .setStyle(2),
-                        new ButtonBuilder()
-                            .setCustomId(`leaderboard_next-${leaderboardPage}`)
-                            .setLabel('Next')
-                            .setDisabled(leaderboard.length < 10)
-                            .setStyle(2)
-                    );
-
-                await interaction.update({embeds: [await sendEmbed(interaction, EmbedType.INFO, `Leaderboard`, `${leaderboardString}`, true, [leaderboardNavigation])]});
-            }
-        });
-    }
-}
-
