+++ src/functions/handlers/commands.js
@@ -3,8 +3,12 @@ const {logMessage} = require('../utilities/core/loggingUtils');
 logMessage(`Hello, world! From handleCommands.js`, `INFO`);
 
 const {REST, Routes, EmbedBuilder} = require('discord.js');
+const guildModel = require("../../models/guildModel");
+const userModel = require("../../models/userModel");
 const token = process.env.TOKEN, clientId = process.env.CLIENT_ID;
 const rest = new REST({version: '9'}).setToken(token);
+
+require('dotenv').config();
 
 /**
  * @name registerSlashCommands
@@ -58,6 +62,37 @@ async function runCode(language, code) {
 
 }
 
+async function createUserIfNotFound(interaction) {
+    const userModel = require('../../models/userModel');
+
+    try {
+        await userModel.create({
+            userID: interaction.user.id,
+            guildID: interaction.guild.id
+        });
+    } catch (error) {
+        logMessage(`Error creating user: ${error.stack}`, 'ERROR');
+        return interaction.reply({ content: 'Whoops! Something went wrong.', ephemeral: true });
+    }
+}
+
+async function createGuildIfNotFound(interaction) {
+    const guildModel = require('../../models/guildModel');
+
+    try {
+        await guildModel.create({
+            guildID: interaction.guild.id,
+            welcomeChannel: interaction.guild.systemChannelId,
+            welcomeMessage: 'Welcome to the server, {{user}}!',
+            leaveChannel: interaction.guild.systemChannelId,
+            leaveMessage: '{{user}} has left the server.'
+        });
+    } catch (error) {
+        logMessage(`Error creating guild: ${error.stack}`, 'ERROR');
+        return interaction.reply({ content: 'Whoops! Something went wrong.', ephemeral: true });
+    }
+}
+
 module.exports = async (client) => {
 
     const {readdirSync} = require('fs');
@@ -68,6 +103,7 @@ module.exports = async (client) => {
         const commands = readdirSync(join(__dirname, '../../commands', dir)).filter(file => file.endsWith('.js'));
         for (const file of commands) {
             const command = require(`../../commands/${dir}/${file}`);
+            logMessage(`    Registered command ${command.data.name}`, `INFO`);
             await client.commands.set(command.data.name, command);
             client.commandArray.push(command.data.toJSON());
         }
@@ -75,5 +111,82 @@ module.exports = async (client) => {
 
     await registerSlashCommands(client, client.commands.map(command => command.data.toJSON()));
     logMessage(`Registered slash commands`, `INFO`);
+
+    client.on('interactionCreate', async interaction => {
+
+        if (interaction.user.bot) return;
+
+        if (interaction.isChatInputCommand()) {
+            logMessage('Command interaction received', 'INFO');
+
+            // Check if X exists
+            const guild = await guildModel.findOne({guildID: interaction.guild.id});
+            if (!guild) await createGuildIfNotFound(interaction);
+
+            const user = await userModel.findOne({userID: interaction.user.id, guildID: interaction.guild.id});
+            if (!user) await createUserIfNotFound(interaction);
+
+            const command = client.commands.get(interaction.commandName);
+            if (!command) return interaction.reply({content: 'Whoops! That command doesn\'t seem to exist.', ephemeral: true});
+
+            try {
+                logMessage(`    Running command ${command.data.name}`, 'INFO');
+                await command.execute(interaction, client);
+            } catch (error) {
+                logMessage(`    Error running command ${command.data.name}`, 'ERROR');
+                logMessage(`    ${error.stack}`, 'ERROR');
+                await interaction.reply({content: 'Whoops! Something went wrong.', ephemeral: true});
+            }
+        }
+        else if (interaction.isModalSubmit()) {
+            const id = interaction.customId;
+
+            if (!id.startsWith('run-code_')) return;
+
+            const language = id.split('_')[1];
+            const code = interaction.fields.getTextInputValue('run-code-input');
+            const output = await runCode(language, code);
+
+            // noinspection JSCheckFunctionSignatures
+            const embed = new EmbedBuilder()
+                .setAuthor({name: 'Made possible by Piston', url: 'https://github.com/engineer-man/piston'})
+                .setTitle(`${output.language} ${output.version} | Your code`)
+                .setDescription(`\`\`\`js\n${code}\`\`\``)
+                .setColor('#00ff00');
+
+            const outputEmbed = new EmbedBuilder()
+                .setTitle('Output')
+                .setDescription(output.run.output || 'No output')
+                .setColor('#00ff00');
+
+            await interaction.reply({embeds: [embed, outputEmbed], ephemeral: false});
+        }
+        else if (interaction.isAutocomplete()) {
+            const command = client.commands.get(interaction.commandName);
+            if (!command) return interaction.reply({content: 'Whoops! Something went wrong.', ephemeral: true});
+
+            try {
+                await command.autocomplete(interaction, client);
+            } catch (error) {
+                await interaction.reply({content: 'Whoops! Something went wrong.', ephemeral: true});
+            }
+        }
+        else if (interaction.isButton()) {
+            const button = client.buttons.get(interaction.customId);
+            if (!button) return interaction.reply({ content: 'Whoops! Something went wrong.', ephemeral: true });
+
+            try {
+                logMessage(`Running button ${button.name}`, `INFO`)
+                await button.execute(interaction, client);
+            } catch (error) {
+                logMessage(`Error running button ${button.name}: ${error}`, `ERROR`)
+                await interaction.reply({ content: 'Whoops! Something went wrong.', ephemeral: true });
+            }
+        }
+        else {
+            logMessage(`Unhandled interaction type: ${interaction.type}`, 'WARNING');
+            return interaction.reply({content: 'Whoops! Something went wrong.', ephemeral: true});
+        }
+    });
 }
